+++ backend/src/main/java/ru/hh/school/service/Vacancy/FavoriteVacancyService.java
@@ -1,63 +1,62 @@
 package ru.hh.school.service.Vacancy;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import jnr.ffi.annotations.In;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.hh.nab.common.properties.FileSettings;
+import ru.hh.school.dao.EmployerDao;
 import ru.hh.school.dao.VacancyDao;
+import ru.hh.school.dto.Area;
+import ru.hh.school.dto.Employer.EmployerShortDto;
+import ru.hh.school.dto.Salary;
 import ru.hh.school.dto.Vacancy.FavoriteVacancyDto;
 import ru.hh.school.dto.Vacancy.VacancyDto;
+import ru.hh.school.entity.FavoriteEmployerEntity;
 import ru.hh.school.entity.FavoriteVacancyEntity;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+
 
 @Service
 public class FavoriteVacancyService {
 
     private final VacancyDao vacancyDao;
+    private final EmployerDao employerDao;
     private final VacancyService vacancyService;
-    private Integer popular_rate;
+    private final Integer popular_rate;
 
     @Autowired
-    public FavoriteVacancyService(VacancyDao vacancyDao, VacancyService vacancyService, FileSettings fileSettings) {
+    public FavoriteVacancyService(VacancyDao vacancyDao, EmployerDao employerDao, VacancyService vacancyService, FileSettings fileSettings) {
         this.vacancyDao = vacancyDao;
+        this.employerDao = employerDao;
         this.vacancyService = vacancyService;
         this.popular_rate = fileSettings.getInteger("popular_rate");
     }
 
     @Transactional
-    public Integer addFavoriteVacancy (Integer vacancyId, String comment) throws JsonProcessingException {
-        VacancyDto vacancyDto = vacancyService.getVacancy(vacancyId);
-        LocalDateTime localDateTime = LocalDateTime.now();
-        Integer views_count  = 0;
-        Integer areaId = Integer.parseInt(vacancyDto.getArea().get("id"));
-        String areaName = vacancyDto.getArea().get("name");
-        Integer salaryFrom = (Integer) vacancyDto.getSalary().get("from");
-        Integer salaryTo = (Integer) vacancyDto.getSalary().get("to");
-        String salaryCurrency = (String) vacancyDto.getSalary().get("currency");
-        Boolean salaryGross = (Boolean) vacancyDto.getSalary().get("gross");
-        Integer employer = vacancyDto.getEmployer();
+    public void  addFavoriteVacancy (String vacancyId, String comment) throws JsonProcessingException {
+        FavoriteVacancyEntity vacancy = vacancyDao.getFavoriteVacancy(vacancyId);
+        if (vacancy == null) {
+            VacancyDto vacancyDto = vacancyService.getVacancy(vacancyId);
 
-        FavoriteVacancyEntity favoriteVacancyEntity = new FavoriteVacancyEntity(
-                vacancyDto.getId(),
-                vacancyDto.getName(),
-                localDateTime,
-                areaId,
-                areaName,
-                salaryFrom,
-                salaryTo,
-                salaryCurrency,
-                salaryGross,
-                vacancyDto.getCreated_at(),
-                comment,
-                employer,
-                views_count);
+            FavoriteVacancyEntity favoriteVacancyEntity = new FavoriteVacancyEntity(
+                    vacancyDto.getId(),
+                    vacancyDto.getName(),
+                    vacancyDto.getArea().getId(),
+                    vacancyDto.getArea().getName(),
+                    vacancyDto.getSalary().getFrom(),
+                    vacancyDto.getSalary().getTo(),
+                    vacancyDto.getSalary().getCurrency(),
+                    vacancyDto.getSalary().getGross(),
+                    vacancyDto.getCreated_at(),
+                    comment,
+                    vacancyDto.getEmployer().getId(),
+                    vacancyDto.getEmployer().getName());
 
-        return vacancyDao.addFavoriteVacancy(favoriteVacancyEntity);
+            vacancyDao.SaveFavoriteVacancy(favoriteVacancyEntity);
+        }
     }
 
     @Transactional
@@ -68,50 +67,67 @@ public class FavoriteVacancyService {
 
         for (FavoriteVacancyEntity vacancy : vacancyDaoList ) {
 
-            favoriteVacancyDto.add(new FavoriteVacancyDto(
+            Area area = new Area(
                     vacancy.getId(),
-                    vacancy.getName(),
-                    vacancy.getData_create(),
-                    vacancy.getAreaId(),
-                    vacancy.getAreaName(),
+                    vacancy.getName());
+
+            Salary salary = new Salary(
                     vacancy.getSalary_from(),
                     vacancy.getSalary_to(),
                     vacancy.getSalary_currency(),
-                    vacancy.getSalary_gross(),
+                    vacancy.getSalary_gross());
+
+            EmployerShortDto employerShortDto = new EmployerShortDto(
+                    vacancy.getEmployer_id(),
+                    vacancy.getEmployer_name());
+
+           favoriteVacancyDto.add(new FavoriteVacancyDto(
+                    vacancy.getId(),
+                    vacancy.getName(),
+                    vacancy.getData_create(),
+                    area,
+                    salary,
                     vacancy.getCreated_at(),
-                    vacancy.getEmployer(),
+                    employerShortDto,
                     vacancy.getViews_count(),
                     vacancy.getComment(),
                     popular_rate));
 
-            vacancy.setViews_count(vacancy.getViews_count()+1);
-            vacancyDao.addFavoriteVacancy(vacancy);
-        }
+            vacancy.setViews_count(vacancy.getViews_count() + 1);
 
+            FavoriteEmployerEntity favoriteEmployerEntity = employerDao.getFavoriteEmployer(vacancy.getEmployer_id());
+
+            if (favoriteEmployerEntity != null)
+            {
+                favoriteEmployerEntity.setViews_count(favoriteEmployerEntity.getViews_count() + 1);
+            }
+        }
         return favoriteVacancyDto;
     }
 
     @Transactional
-    public void deleteFavoriteVacancy(Integer vacancyId) {
-        vacancyDao.deleteFavoriteVacancy(vacancyId);
+    public void deleteFavoriteVacancy(String vacancyId) {
+        FavoriteVacancyEntity vacancy = vacancyDao.getFavoriteVacancy(vacancyId);
+        if (vacancy != null) {
+            vacancyDao.deleteFavoriteVacancy(vacancyId);
+        }
     }
 
     @Transactional
-    public void refreshFavoriteVacancy(Integer vacancyId) throws JsonProcessingException {
+    public void refreshFavoriteVacancy(String vacancyId) throws JsonProcessingException {
         VacancyDto vacancyDto = vacancyService.getVacancy(vacancyId);
-        FavoriteVacancyEntity vacancy = vacancyDao.getFavoriteVacancy(vacancyId)
-                .orElseThrow(() -> new IllegalArgumentException("Vacancy not found with id " + vacancyId));
+        FavoriteVacancyEntity vacancy = vacancyDao.getFavoriteVacancy(vacancyId);
 
-        vacancy.setName(vacancyDto.getName());
-
-        vacancy.setSalary_from(Integer.parseInt((String) vacancyDto.getSalary().get("from")));
-        vacancy.setSalary_to(Integer.parseInt((String) vacancyDto.getSalary().get("to")));
-        vacancy.setSalary_currency((String) vacancyDto.getSalary().get("currency"));
-        vacancy.setSalary_gross((Boolean) vacancyDto.getSalary().get("gross"));
-
-        vacancy.setAreaId(Integer.parseInt(vacancyDto.getArea().get("id")));
-        vacancy.setAreaName(vacancyDto.getArea().get("name"));
-
-        vacancyDao.addFavoriteVacancy(vacancy);
+        if (vacancy != null)
+        {
+            vacancy.setName(vacancyDto.getName());
+            vacancy.setAreaId(vacancyDto.getArea().getId());
+            vacancy.setAreaName(vacancyDto.getArea().getName());
+            vacancy.setSalary_from(vacancyDto.getSalary().getFrom());
+            vacancy.setSalary_to(vacancyDto.getSalary().getTo());
+            vacancy.setSalary_currency(vacancyDto.getSalary().getCurrency());
+            vacancy.setSalary_gross(vacancyDto.getSalary().getGross());
+        }
     }
+
 }
